creating test cases for filters and beatpair

+++ R peak detection/beatdetection/ptemplate.py
@@ -12,7 +12,7 @@ p_wave = []
 
 def qrs(
         r: np.ndarray,
-        n: int) -> int:
+        n: int) -> np.int64:
 
     """
